2.11 [21:30] verify cycle & int -> size_t

+++ include/list_manager.h
@@ -41,8 +41,8 @@ typedef struct
     FILE* log_file;
 } List_t;
 
-int         ListHead        (List_t* list);
-int         ListTail        (List_t* list);
+size_t      ListHead        (List_t* list);
+size_t      ListTail        (List_t* list);
 ListErr     ListAddFront    (List_t* list, list_type value);
 ListErr     ListAddBack     (List_t* list, list_type value);
 list_type   ListGetOnIndex  (List_t* list, int index);
@@ -53,8 +53,8 @@ void        ListDestroy     (List_t* list);
 ListErr     ListAddAfter    (List_t* list, int index, list_type value);
 ListErr     ListAddBefore   (List_t* list, int index, list_type value);
 ListErr     ListDel         (List_t* list, int index);
-int         ListNext        (List_t* list, int index);
-int         ListPrev        (List_t* list, int index);
+size_t      ListNext        (List_t* list, int index);
+size_t      ListPrev        (List_t* list, int index);
 ListErr     ListVerify      (List_t* list);
 void PrintError(ListErr errcode);
 int ValueEquality(list_type a, list_type b);
@@ -65,21 +65,21 @@ const size_t LIST_MULTIPLIER_CAPACITY = 2;
 const list_type SHIELD_VALUE = 0xF0F0F0F0;
 const list_type POISON = 0xBADF00D;
 
-#define CHECK(errcode)\
-if(errcode != LIST_CORRECT)\
-{\
-    PrintError(errcode);\
-    return errcode;\
+#define CHECK(errcode)          \
+if(errcode != LIST_CORRECT)     \
+{                               \
+    PrintError(errcode);        \
+    return errcode;             \
 }
-#define VERIFY(list) \
-{\
-    ListErr errr = ListVerify(list);\
-    if(errr != LIST_CORRECT)\
-    {\
-        PrintError(errr);\
-        ListDump(list);\
-        return errr;\
-    }\
+#define VERIFY(list)                    \
+{                                       \
+    ListErr errr = ListVerify(list);    \
+    if(errr != LIST_CORRECT)            \
+    {                                   \
+        PrintError(errr);               \
+        ListDump(list);                 \
+        return errr;                    \
+    }                                   \
 }
 
 #endif // LIST_MANAGER_H_
